Delete components when cluster is deleted
- Do not delete portworx storage classes if there is
  no delete strategy. As portworx is not cleaned up,
  there could be volumes that are using the storage class.
- Delete node wiper as well when finished running

+++ drivers/storage/portworx/component/portworx_sc.go
@@ -186,6 +186,13 @@ annotations:
 }
 
 func (c *portworxStorageClass) Delete(cluster *corev1alpha1.StorageCluster) error {
+	if cluster.DeletionTimestamp != nil &&
+		(cluster.Spec.DeleteStrategy == nil || cluster.Spec.DeleteStrategy.Type == "") {
+		// If the cluster is deleted without any delete strategy do not delete the
+		// storage classes as Portworx is still running on the nodes
+		return nil
+	}
+
 	if err := k8sutil.DeleteStorageClass(c.k8sClient, PxDbStorageClass); err != nil {
 		return err
 	}
